Grade activity working. Converts grade to letter and back. Refactored sharedPreference to helper class.

+++ app/src/main/java/com/example/gradeviewer/Assignment.java
@@ -5,10 +5,10 @@ import java.util.Random;
 public class Assignment {
     private static int assID = 0; //static ID increments with every new assignment created
     private String assignmentTitle; //title of assignment
-    private int assignmentGrade; //grade of assignment
+    private double assignmentGrade; //grade of assignment
 
     //private constructor. Increments ID.
-    private Assignment(String title, int grade){
+    private Assignment(String title, double grade){
         assignmentTitle = title;
         assignmentGrade = grade;
         assID++;
@@ -18,11 +18,11 @@ public class Assignment {
     static public Assignment generateRandomAssignment(){
         Random rnd = new Random();
         String tempTitle = "Assignment " + assID;
-        int tempGrade = rnd.nextInt(100) + 1;
+        double tempGrade = rnd.nextInt(100) + 1;
         return new Assignment(tempTitle, tempGrade);
     }
 
     // Getters
     public String getAssignmentTitle(){ return assignmentTitle; }
-    public int getAssignmentGrade(){ return assignmentGrade; }
+    public double getAssignmentGrade(){ return assignmentGrade; }
 }
